// src/front/src/api/identity/models/appUserResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX Identity API (via Gateway)
 * OpenAPI spec version: 1.0
 */
import type { Department } from "./department";

export interface AppUserResponse {
	id?: number;
	/** @nullable */
	gid?: string | null;
	/** @nullable */
	firstName?: string | null;
	/** @nullable */
	lastName?: string | null;
	/** @nullable */
	email?: string | null;
	department?: Department;
	structureId?: number;
	/** @nullable */
	teamId?: number | null;
	/** @nullable */
	supervisorId?: number | null;
	/** @nullable */
	supervisorFullName?: string | null;
	isActive?: boolean;
	/** @nullable */
	roles?: string[] | null;
	createdAt?: string;
	/** @nullable */
	modifiedAt?: string | null;
}